Tests/requirements.txt
diffusers
invisible_watermark
transformers
accelerate
safetensors
python-slugify
scipy
timm
torchvision
torchaudio